refactor: перемещает usePageVisibility в services/hooks
